refactor: Address PR review feedback
- #4 (mandatory): Extract `QUOTE_WRAPPER_CHARS` constant in
  utils/generic.ts. `stripQuoteWrappers` and `normalizeRecordKey`
  now derive their regexes from one source — adding a new wrapper
  char (e.g. «») updates both call sites in one place.
- #1: Drop `normalizeStorageId`. It was a one-line delegation to
  `stripQuoteWrappers`; 6 call sites now use the underlying helper
  directly (no premature abstraction).
- #2: Move `wrapJsonText` from `tools/common/storage_helpers.ts` to
  `utils/mcp.ts`. It's response-formatting, not storage-specific;
  it now sits next to `buildMCPResponse`. 8 tool files updated.
- #3: Stop exporting `parseFencedJson` from production code. The
  test helper at `tests/unit/helpers/tool_context.ts` now owns its
  own implementation that imports the `JSON_FENCE_PREFIX` /
  `JSON_FENCE_SUFFIX` constants from `utils/mcp.ts`, so prod and
  tests can't drift on the fence shape.

#5 (co-locate response builders) is tracked by #929.
