+++ src/tools/common/get_key_value_store_keys.ts
@@ -5,8 +5,10 @@ import { HelperTools, HTTP_NOT_FOUND } from '../../const.js';
 import type { InternalToolArgs, ToolEntry, ToolInputSchema } from '../../types.js';
 import { TOOL_TYPE } from '../../types.js';
 import { compileSchema } from '../../utils/ajv.js';
+import { stripQuoteWrappers } from '../../utils/generic.js';
 import { getHttpStatusCode } from '../../utils/logging.js';
-import { buildStorageNotFound, normalizeStorageId, wrapJsonText } from './storage_helpers.js';
+import { wrapJsonText } from '../../utils/mcp.js';
+import { buildStorageNotFound } from './storage_helpers.js';
 
 const getKeyValueStoreKeysArgs = z.object({
     keyValueStoreId: z.string().min(1).describe('Key-value store ID or username~store-name'),
@@ -47,7 +49,7 @@ export const getKeyValueStoreKeys: ToolEntry = Object.freeze({
     call: async (toolArgs: InternalToolArgs) => {
         const { args, apifyClient: client } = toolArgs;
         const parsed = getKeyValueStoreKeysArgs.parse(args);
-        const keyValueStoreId = normalizeStorageId(parsed.keyValueStoreId);
+        const keyValueStoreId = stripQuoteWrappers(parsed.keyValueStoreId);
         // `listKeys()` throws ApifyApiError on a missing store (the SDK only soft-catches
         // 404 on `.get()` / `.getRecord()`), so translate 404 into a soft-fail.
         const keys = await client
