+++ src/tools/common/get_key_value_store_record.ts
@@ -5,7 +5,9 @@ import { HelperTools } from '../../const.js';
 import type { InternalToolArgs, ToolEntry, ToolInputSchema } from '../../types.js';
 import { TOOL_TYPE } from '../../types.js';
 import { compileSchema } from '../../utils/ajv.js';
-import { buildStorageNotFound, normalizeRecordKey, normalizeStorageId, wrapJsonText } from './storage_helpers.js';
+import { stripQuoteWrappers } from '../../utils/generic.js';
+import { wrapJsonText } from '../../utils/mcp.js';
+import { buildStorageNotFound, normalizeRecordKey } from './storage_helpers.js';
 
 const getKeyValueStoreRecordArgs = z.object({
     keyValueStoreId: z.string().min(1).describe('Key-value store ID or username~store-name'),
@@ -41,7 +43,7 @@ export const getKeyValueStoreRecord: ToolEntry = Object.freeze({
     call: async (toolArgs: InternalToolArgs) => {
         const { args, apifyClient: client } = toolArgs;
         const parsed = getKeyValueStoreRecordArgs.parse(args);
-        const keyValueStoreId = normalizeStorageId(parsed.keyValueStoreId);
+        const keyValueStoreId = stripQuoteWrappers(parsed.keyValueStoreId);
         const recordKey = normalizeRecordKey(parsed.recordKey);
         const store = client.keyValueStore(keyValueStoreId);
         const record = await store.getRecord(recordKey);
