+++ tests/unit/tools.storage_helpers.test.ts
@@ -1,27 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
 import { FAILURE_CATEGORY, TOOL_STATUS } from '../../src/const.js';
-import {
-    buildStorageNotFound,
-    normalizeRecordKey,
-    normalizeStorageId,
-    parseFencedJson,
-    wrapJsonText,
-} from '../../src/tools/common/storage_helpers.js';
-
-describe('wrapJsonText()', () => {
-    it('emits a ```json … ``` fenced block', () => {
-        expect(wrapJsonText({ a: 1 })).toBe('```json\n{"a":1}\n```');
-    });
-
-    it('serializes arrays', () => {
-        expect(wrapJsonText([1, 2])).toBe('```json\n[1,2]\n```');
-    });
-
-    it('serializes primitives', () => {
-        expect(wrapJsonText('x')).toBe('```json\n"x"\n```');
-    });
-});
+import { buildStorageNotFound, normalizeRecordKey } from '../../src/tools/common/storage_helpers.js';
 
 describe('buildStorageNotFound()', () => {
     it('returns a SOFT_FAIL / INVALID_INPUT response with the supplied message', () => {
@@ -33,13 +13,6 @@ describe('buildStorageNotFound()', () => {
             toolStatus: TOOL_STATUS.SOFT_FAIL,
             failureCategory: FAILURE_CATEGORY.INVALID_INPUT,
         });
-    });
-});
-
-describe('normalizeStorageId()', () => {
-    // Wrapper-stripping behavior is pinned via `stripQuoteWrappers` in utils.generic.test.ts.
-    it('delegates to stripQuoteWrappers — typical wrapped id is returned canonical', () => {
-        expect(normalizeStorageId('`user~my-store`')).toBe('user~my-store');
     });
 });
 
@@ -57,13 +30,5 @@ describe('normalizeRecordKey()', () => {
 
     it('trims surrounding whitespace', () => {
         expect(normalizeRecordKey('  INPUT  ')).toBe('INPUT');
-    });
-});
-
-describe('parseFencedJson()', () => {
-    it('round-trips with wrapJsonText for objects, arrays, and primitives', () => {
-        for (const value of [{ a: 1, b: [2, 3] }, [1, 2, 3], 'x', 42, true, null]) {
-            expect(parseFencedJson(wrapJsonText(value))).toEqual(value);
-        }
     });
 });
