+++ tests/unit/utils.mcp.test.ts
@@ -1,6 +1,20 @@
 import { describe, expect, it } from 'vitest';
 
-import { computeToolResponseSizeBytes } from '../../src/utils/mcp.js';
+import { computeToolResponseSizeBytes, wrapJsonText } from '../../src/utils/mcp.js';
+
+describe('wrapJsonText()', () => {
+    it('emits a ```json … ``` fenced block', () => {
+        expect(wrapJsonText({ a: 1 })).toBe('```json\n{"a":1}\n```');
+    });
+
+    it('serializes arrays', () => {
+        expect(wrapJsonText([1, 2])).toBe('```json\n[1,2]\n```');
+    });
+
+    it('serializes primitives', () => {
+        expect(wrapJsonText('x')).toBe('```json\n"x"\n```');
+    });
+});
 
 describe('computeToolResponseSizeBytes()', () => {
     it('returns 0 for null/undefined/non-object input', () => {
